% tinyrox says don't edit this manually, but it can't stop you!
\name{check_setup}
\alias{check_setup}
\title{Quick check that repo is wired correctly}
\usage{
check_setup(repo_dir = getOption("hacer.repo", getwd()))
}
\arguments{
\item{repo_dir}{Path to your ToDo repo directory.}
}
\description{
Quick check that repo is wired correctly
}
