% tinyrox says don't edit this manually, but it can't stop you!
\name{infer_period_from_filename}
\alias{infer_period_from_filename}
\title{Infer a period name from a ToDo filename}
\usage{
infer_period_from_filename(f)
}
\arguments{
\item{f}{A filename or path like `ToDo_250915_Daily.txt`.}
}
\description{
Infer a period name from a ToDo filename
}
