% tinyrox says don't edit this manually, but it can't stop you!
\name{inherit_recur_to_parents}
\alias{inherit_recur_to_parents}
\title{Bubble `recur = TRUE` up to ancestor tasks}
\usage{
inherit_recur_to_parents(df)
}
\arguments{
\item{df}{A parsed ToDo data.frame.}
}
\description{
Bubble `recur = TRUE` up to ancestor tasks
}
