% tinyrox says don't edit this manually, but it can't stop you!
\name{instantiate_todo}
\alias{instantiate_todo}
\title{Create a hacer project layout in a Git repo.}
\usage{
instantiate_todo(repo_dir, syncthing_live_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{repo_dir}{Path to your ToDo repo directory.}

\item{syncthing_live_dir}{Optional path to a Syncthing-shared live directory.}

\item{overwrite}{Overwrite existing files? Default `FALSE`.}
}
\description{
Create a hacer project layout in a Git repo.
}
