% tinyrox says don't edit this manually, but it can't stop you!
\name{open_this_week}
\alias{open_this_week}
\title{Open this week's files for quick editing}
\usage{
open_this_week(date = Sys.Date(), cfg = todo_config())
}
\arguments{
\item{date}{A Date. Defaults to `Sys.Date()`.}

\item{cfg}{A config list from `todo_config()`.}
}
\description{
Open this week's files for quick editing
}
