% tinyrox says don't edit this manually, but it can't stop you!
\name{print.hacer_preview}
\alias{print.hacer_preview}
\title{Print a hacer_preview summary}
\usage{
\method{print}{hacer_preview}(x, ...)
}
\arguments{
\item{x}{A `hacer_preview` object.}

\item{...}{Unused.}
}
\description{
Print a hacer_preview summary
}
