% tinyrox says don't edit this manually, but it can't stop you!
\name{propagate_from_daily}
\alias{propagate_from_daily}
\title{Propagate status from Daily into a target period}
\usage{
propagate_from_daily(daily, target)
}
\arguments{
\item{daily}{A parsed Daily data.frame.}

\item{target}{A parsed target data.frame (Week, Month, or Quarter).}
}
\description{
Propagate status from Daily into a target period
}
