% tinyrox says don't edit this manually, but it can't stop you!
\name{fix_parents}
\alias{fix_parents}
\title{Roll up parent statuses in a single file}
\usage{
fix_parents(file_name)
}
\arguments{
\item{file_name}{Path to a ToDo `.txt` file.}
}
\description{
Roll up parent statuses in a single file
}
