% tinyrox says don't edit this manually, but it can't stop you!
\name{sync_from_daily}
\alias{sync_from_daily}
\title{Sync new items added in Daily up to Week/Month/Quarter}
\usage{
sync_from_daily(date = Sys.Date(), cfg = todo_config())
}
\arguments{
\item{date}{A Date. Defaults to `Sys.Date()`.}

\item{cfg}{A config list from `todo_config()`.}
}
\description{
Sync new items added in Daily up to Week/Month/Quarter
}
