% tinyrox says don't edit this manually, but it can't stop you!
\name{todo_config}
\alias{todo_config}
\title{Load the ToDo repo configuration}
\usage{
todo_config(repo_dir = getOption("hacer.repo", getwd()))
}
\arguments{
\item{repo_dir}{Path to your ToDo repo directory.}
}
\description{
Load the ToDo repo configuration
}
